# tests/fixtures/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_AlertsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[AlertsData]",
        "type": "object"
      },
      "Envelope_InboundData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InboundData"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[InboundData]",
        "type": "object"
      },
      "Envelope_PlaceData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceData"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PlaceData]",
        "type": "object"
      },
      "Envelope_RecommendationsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RecommendationsData]",
        "type": "object"
      },
...
      "Envelope_VisitorForecastData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VisitorForecastData"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[VisitorForecastData]",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "RecommendationRequest": {
        "additionalProperties": false,
        "properties": {
          "area_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Code"
          },
          "budget_krw": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Krw"
          },
          "constraints": {
            "$ref": "#/components/schemas/RecommendationConstraints"
          },
          "limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "party_size": {
            "default": 1,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Party Size",
            "type": "integer"
          },
          "target_country": {
            "pattern": "^[A-Za-z]{2}$",
            "title": "Target Country",
            "type": "string"
          },
          "themes": {
            "items": {
              "enum": [
                "nature",
                "culture",
                "food",
                "kpop"
              ],
              "type": "string"
            },
            "title": "Themes",
            "type": "array"
          },
          "travel_window": {
            "$ref": "#/components/schemas/TravelWindow"
          }
        },
        "required": [
          "target_country",
          "travel_window"
        ],
        "title": "RecommendationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "EDEN DB에 주기적으로 수집하고 계산한 현재 상태만 반환합니다. API 요청은 외부 데이터 수집을 유발하지 않습니다. 관광지 content_id는 EDEN ID가 기준이며 기존 TourAPI content ID도 별칭으로 조회할 수 있습니다.",
    "summary": "한국 관광 트렌드와 방한시장 데이터 API",
    "title": "EDEN API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/forecasts/visitors": {
      "get": {
        "operationId": "get_visitor_forecast_v1_forecasts_visitors_get",
        "parameters": [
          {
            "in": "query",
            "name": "area_code",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Area Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "place_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 300,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Place Name"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "nx",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nx"
            }
          },
          {
            "in": "query",
            "name": "ny",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ny"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "weather",
                      "festivals",
                      "holidays"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_VisitorForecastData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Visitor Forecast",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/v1/markets/inbound": {
      "get": {
        "operationId": "get_inbound_markets_v1_markets_inbound_get",
        "parameters": [
          {
            "in": "query",
            "name": "countries",
            "required": true,
            "schema": {
              "items": {
                "pattern": "^[A-Za-z]{2}$",
                "type": "string"
              },
              "minItems": 1,
              "title": "Countries",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "social_sources",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "youtube",
                      "instagram",
                      "tiktok",
                      "x",
                      "reddit",
                      "weibo",
                      "douyin",
                      "xiaohongshu",
                      "line",
                      "facebook"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Sources"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "12m",
              "enum": [
                "3m",
                "6m",
                "12m",
                "24m"
              ],
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          },
          {
            "in": "query",
            "name": "forecast_days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 7,
              "minimum": 1,
              "title": "Forecast Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "visitors",
                      "flights",
                      "flight_schedule",
                      "fx",
                      "tourism_balance",
                      "social_interest"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_InboundData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Inbound Markets",
        "tags": [
          "markets"
        ]
      }
    },
    "/v1/markets/{country}/alerts": {
      "get": {
        "operationId": "get_market_alerts_v1_markets__country__alerts_get",
        "parameters": [
          {
            "in": "path",
            "name": "country",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z]{2}$",
              "title": "Country",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "visa",
                      "entry",
                      "safety",
                      "travel",
                      "market_trend"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Types"
            }
          },
          {
            "in": "query",
            "name": "source_scope",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "korean",
                "local",
                "all"
              ],
              "title": "Source Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "ko",
              "enum": [
                "ko",
                "en"
              ],
              "title": "Language",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AlertsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Market Alerts",
        "tags": [
          "markets"
        ]
      }
    },
    "/v1/places/{content_id}": {
      "get": {
        "operationId": "get_place_v1_places__content_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Content Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ko",
              "enum": [
                "ko",
                "en",
                "ja",
                "zh-CN"
              ],
              "title": "Lang",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "radius_m",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 5000,
              "minimum": 100,
              "title": "Radius M",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "related_limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 1,
              "title": "Related Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "related",
                      "shops",
                      "hub"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlaceData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Place",
        "tags": [
          "places"
        ]
      }
    },
    "/v1/recommendations/destinations": {
      "post": {
        "operationId": "recommend_destinations_v1_recommendations_destinations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RecommendationsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Recommend Destinations",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}